+++ src/main/java/de/kamillionlabs/hateoflux/assembler/SealedResourceListAssemblerModule.java
@@ -61,11 +61,17 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
     }
 
     /**
-     * Creates an empty {@link HalListWrapper} including hypermedia links applicable to the entire list.
+     * Creates a paged empty {@link HalListWrapper} including hypermedia links applicable to the entire list.
      *
      * @param listItemTypeAsNameOrigin
      *         the class of the resource from which the list name is derived, typically pluralized to represent both
      *         resource and embedded resource types (also see {@link Relation})
+     * @param pageSize
+     *         the requested/max number of elements in a single page
+     * @param pageNumber
+     *         the current page number
+     * @param sortCriteria
+     *         sort criteria (property and direction) of the page
      * @param exchange
      *         provides the context of the current web exchange, such as the base URL
      * @return an initialized {@link HalListWrapper} with relevant hypermedia links for the entire list
@@ -74,9 +80,13 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
      */
     default HalListWrapper<ResourceT, EmbeddedT> createEmptyListWrapper(@NonNull Class<?> listItemTypeAsNameOrigin,
                                                                         int pageSize,
+                                                                        int pageNumber,
+                                                                        List<SortCriteria> sortCriteria,
                                                                         ServerWebExchange exchange) {
         HalListWrapper<ResourceT, EmbeddedT> emptyWrapper = HalListWrapper.empty(listItemTypeAsNameOrigin);
-        return emptyWrapper.withLinks(buildLinksForResourceList(exchange));
+        HalPageInfo halPageInfo = HalPageInfo.assembleWithPageNumber(pageSize, 0, pageNumber);
+        return emptyWrapper.withLinks(buildLinksForResourceList(halPageInfo, sortCriteria, exchange))
+                .withPageInfo(halPageInfo);
     }
 
     /**
@@ -101,6 +111,12 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
      *
      * @param listName
      *         the given name for the list
+     * @param pageSize
+     *         the requested/max number of elements in a single page
+     * @param pageNumber
+     *         the current page number
+     * @param sortCriteria
+     *         sort criteria (property and direction) of the page
      * @param exchange
      *         provides the context of the current web exchange, such as the base URL
      * @return an initialized {@link HalListWrapper} with relevant hypermedia links for the entire list
@@ -108,9 +124,13 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
      * @see #createEmptyListWrapper(Class, ServerWebExchange)
      */
     default HalListWrapper<ResourceT, EmbeddedT> createEmptyListWrapper(@NonNull String listName,
+                                                                        int pageSize,
+                                                                        int pageNumber,
+                                                                        List<SortCriteria> sortCriteria,
                                                                         ServerWebExchange exchange) {
         HalListWrapper<ResourceT, EmbeddedT> emptyWrapper = HalListWrapper.empty(listName);
-        return emptyWrapper.withLinks(buildLinksForResourceList(exchange));
+        HalPageInfo halPageInfo = HalPageInfo.assembleWithPageNumber(pageSize, 0, pageNumber);
+        return emptyWrapper.withLinks(buildLinksForResourceList(halPageInfo, sortCriteria, exchange));
     }
 
     /**
@@ -131,10 +151,14 @@ public sealed interface SealedResourceListAssemblerModule<ResourceT, EmbeddedT> 
     }
 
     /**
-     * Main method for building all links for a list of resources and embedded resources, including a
+     * Main method for building all links for a paged list of resources and embedded resources, including
      * navigational links and other contextual links applicable to the entire list. It aggregates results from
      * {@link #buildSelfLinkForResourceList} and {@link #buildOtherLinksForResourceList}.
      *
+     * @param pageInfo
+     *         pagination information about the request page
+     * @param sortCriteria
+     *         sort criteria (property and direction) of the page
      * @param exchange
      *         provides the context of the current web exchange, such as the base URL
      * @return a list of {@link Link} objects representing hypermedia links for the entire list of resource and embedded